[Merged] Adapted merge from Release
Major:
- The tutorial data of set 1 has had been wrongfully replaced with non-linear synthetic images in while merging a pull request in the Release branch #2
- So, we merged Release in Pre-release while making sure to discard the changes made to the tutorial dataset since the correct data are still on Pre-Release

+++ Segmentation_Otsu/data.py
@@ -142,7 +142,7 @@ class Data:
 
         self.mask_Otsu = o.segmentation_otsu(self.image_ExG, start_threshold)
         #conversion en noir et blanc :
-        self.mask_Otsu=np.where(self.mask_Otsu<=200, 0, 255)
+        self.mask_Otsu=np.where(self.mask_Otsu<=200, 0, 255).astype(np.uint8)
     
        
     def create_maskfusion(self):
